models/plant_disease/requirements.txt
kagglehub
tensorflow[standard]
keras[standard]
numpy
pillow